WasafiRetPr/requirements.txt
Pillow-10.4.0
django-widget-tweaks-1.5.0
python-dotenv-1.0.1

